Fix oracle and signal_registry merge corruption blocking compilation.
Split duplicated template definitions, restore missing modules and event
helpers, and resolve type mismatches introduced by the main branch merge.
