+++ stellar-swipe/contracts/signal_registry/src/lib.rs
@@ -18,6 +18,7 @@ mod migration;
 mod ml_scoring;
 mod multisig_approvals;
 mod performance;
+mod providers;
 mod query;
 mod reports;
 pub mod reputation;
@@ -27,6 +28,7 @@ mod social;
 mod stake;
 mod storage_monitor;
 mod submission;
+mod template_presets;
 mod templates;
 mod test_reputation;
 mod types;
@@ -43,7 +45,7 @@ use admin::{
     require_not_paused_legacy as require_not_paused, AdminConfig,
 };
 use shared::version::{set_contract_version, SIGNAL_REGISTRY_VERSION};
-use stellar_swipe_common::emergency::PauseState;
+use stellar_swipe_common::emergency::{PauseState, CAT_SIGNALS, CAT_TRADING};
 use stellar_swipe_common::rate_limit::{self as rl, ActionType as RLAction, RateLimitConfig};
 use stellar_swipe_common::SECONDS_PER_30_DAY_MONTH;
 
@@ -66,20 +68,25 @@ pub use leaderboard::{
     ProviderLeaderboard, ProviderLeaderboardEntry, ProviderMetric,
 };
 pub use ml_scoring::{MLModel, SignalFeatures, SignalScore};
+use providers::VerificationEligibility;
 use reputation::{
     calculate_trust_score, get_trust_score, update_median_values, update_trust_score,
     TrustScoreDetails, TrustScoreTier,
 };
-use soroban_sdk::{contract, contractimpl, contracttype, Address, Bytes, Env, Map, String, Vec};
+use soroban_sdk::{
+    contract, contractimpl, contracttype, Address, Bytes, Env, IntoVal, Map, String, Symbol, Val,
+    Vec,
+};
 use stellar_swipe_common::{health_uninitialized, placeholder_admin, HealthStatus};
 use stellar_swipe_common::{validate_asset_pair as validate_asset_pair_common, AssetPairError};
 use stellar_swipe_common::{ApprovalProposal, MultisigTimelockConfig, ProposalStatus};
-pub use templates::{SignalTemplate, SignalTemplateOverrides, StoredSignalTemplate};
-use templates::{SignalTemplate, DEFAULT_TEMPLATE_EXPIRY_HOURS};
+pub use template_presets::{SignalTemplateOverrides, SignalTemplatePreset, StoredSignalTemplate};
+pub use templates::SignalTemplate;
+use templates::DEFAULT_TEMPLATE_EXPIRY_HOURS;
 use types::{
-    AddressMapping, Asset, CrossChainSignal, FeeBreakdown, ImportResultView, ProviderMonthlyReport,
-    ProviderPerformance, RecurrencePattern, Signal, SignalData, SignalEditInput, SignalOutcome,
-    SignalPerformanceView, SignalStatus, SignalSummary, SortOption, SyncStatus, TradeExecution,
+    AddressMapping, Asset, CrossChainSignal, ImportResultView, ProviderMonthlyReport,
+    RecurrencePattern, Signal, SignalData, SignalEditInput, SignalPerformanceView, SignalSummary,
+    SortOption, SyncStatus, TradeExecution,
 };
 use versioning::{CopyRecord, SignalVersion};
 
@@ -662,8 +669,9 @@ impl SignalRegistry {
 
     /// Returns `true` if the Stellar account for `provider` still exists on-chain.
     /// A merged (deleted) account returns `false`.
-    fn check_provider_exists(env: &Env, provider: &Address) -> bool {
-        env.accounts().get(provider).is_some()
+    fn check_provider_exists(_env: &Env, _provider: &Address) -> bool {
+        // Account existence is not queryable from contract WASM; assume present.
+        true
     }
 
     /// Mark a signal as orphaned (provider account deleted), emit the event, and persist.
@@ -886,8 +894,8 @@ impl SignalRegistry {
         }
 
         // Check for expiry warning (Issue #417)
-        let current_ledger = env.ledger().sequence();
-        let time_to_expiry = signal.expiry.saturating_sub(current_ledger);
+        let now = env.ledger().timestamp();
+        let time_to_expiry = signal.expiry.saturating_sub(now);
         if time_to_expiry <= WARNING_WINDOW_LEDGERS && !signal.warning_emitted {
             events::emit_signal_expiry_warning(
                 &env,
@@ -907,14 +915,14 @@ impl SignalRegistry {
     pub fn save_signal_template(
         env: Env,
         provider: Address,
-        template: SignalTemplate,
+        template: SignalTemplatePreset,
     ) -> Result<u32, AdminError> {
         provider.require_auth();
         Self::validate_asset_pair(&env, &template.asset_pair)?;
 
         let mut templates_map = Self::get_signal_templates_map(&env);
         let template_id =
-            templates::save_signal_template(&env, &mut templates_map, provider, template)
+            template_presets::save_signal_template(&env, &mut templates_map, provider, template)
                 .map_err(|_| AdminError::InvalidParameter)?;
         Self::save_signal_templates_map(&env, &templates_map);
 
@@ -929,13 +937,25 @@ impl SignalRegistry {
     ) -> Result<u64, AdminError> {
         let templates_map = Self::get_signal_templates_map(&env);
         let template =
-            templates::get_signal_template(&templates_map, provider.clone(), template_id)
+            template_presets::get_signal_template(&templates_map, provider.clone(), template_id)
                 .map_err(|_| AdminError::InvalidParameter)?;
         let (asset_pair, action, expiry_hours, price, rationale) =
-            templates::merge_template(template, overrides);
+            template_presets::merge_template(template, overrides);
         let expiry = env.ledger().timestamp() + expiry_hours * 60 * 60;
+        let tags = Vec::new(&env);
 
-        Self::create_signal(env, provider, asset_pair, action, price, rationale, expiry)
+        Self::create_signal(
+            env,
+            provider,
+            asset_pair,
+            action,
+            price,
+            rationale,
+            expiry,
+            SignalCategory::SWING,
+            tags,
+            RiskLevel::Medium,
+        )
     }
 
     /* =========================
@@ -1479,7 +1499,13 @@ impl SignalRegistry {
         }
 
         let mut stakes = Self::get_provider_stakes_map(&env);
-        stakes.set(provider, amount);
+        let mut info = stakes.get(provider.clone()).unwrap_or(stake::StakeInfo {
+            amount: 0,
+            last_signal_time: 0,
+            locked_until: 0,
+        });
+        info.amount = amount;
+        stakes.set(provider, info);
         Self::save_provider_stakes_map(&env, &stakes);
         Ok(())
     }
@@ -1536,7 +1562,10 @@ impl SignalRegistry {
     pub fn check_verification_eligibility(env: Env, provider: Address) -> VerificationEligibility {
         let stakes = Self::get_provider_stakes_map(&env);
         let stats = Self::get_provider_stats_map(&env);
-        let stake = stakes.get(provider.clone()).unwrap_or(0);
+        let stake = stakes
+            .get(provider.clone())
+            .map(|info| info.amount)
+            .unwrap_or(0);
         let performance = stats.get(provider.clone()).unwrap_or_default();
 
         providers::check_verification_eligibility(&env, provider, stake, performance)
@@ -2752,13 +2781,7 @@ pub struct StorageStats {
 #[cfg(test)]
 mod test;
 #[cfg(test)]
-mod test;
-#[cfg(test)]
 mod test_admin_transfer;
-#[cfg(test)]
-mod test_admin_transfer;
-#[cfg(test)]
-mod test_adoption;
 #[cfg(test)]
 mod test_adoption;
 #[cfg(test)]
